refactor: rename connections to contacts
